# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "idus API",
        "description": "\n ## Request 인증\n - `/api/token` 엔드포인트에서 인증 토큰을 발급 받을 수 있습니다.\n - Request Header `Authorization` 필드 안에 `bearer <auth-token>` 형태로 인증 토큰을 기입 후 요청해주시면 됩니다.\n\n ## 유효성 검사 실패\n - Response Code: 400\n - Response Body: `ResponseError`\n ```json\n {\n    'success': false,\n    'message': '유효성 검사에 실패했습니다.',\n    'errors': {\n        'username': [\n            '이미 존재하는 username 입니다.'\n        ]\n    }\n}\n ```\n",
        "version": "0.1"
    },
    "paths": {
        "/api/register": {
            "post": {
                "tags": [
                    "Login"
                ],
                "summary": "회원 가입",
                "description": "",
                "operationId": "a39949df013c1eb78ff8bd8f481d408a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/NewUser"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "성공",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/token": {
            "post": {
                "tags": [
                    "Login"
                ],
                "summary": "토큰 발급",
                "description": "",
                "operationId": "46986b9904b82938eb2d7656c737fc43",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "username": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "성공",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "allOf": [
                                        {
                                            "properties": {
                                                "result": {
                                                    "properties": {
                                                        "access_token": {
                                                            "description": "API 접근 토큰",
                                                            "example": "1|s3AnvYQ4k4F56JgjNvq163DTDB4csAoDRDknqp8E"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        {
                                            "$ref": "#/components/schemas/ResponseObject"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "ResponseData": {
                "description": "리스폰 데이터",
                "properties": {
                    "success": {
                        "type": "boolean"
                    }
                },
                "type": "object"
            },
            "ResponseError": {
                "description": "리스폰 에러 객체",
                "type": "object",
                "example": {
                    "success": false,
                    "message": "유효성 검사에 실패했습니다.",
                    "errors": {
                        "username": [
                            "username 값이 6 ~ 25 글자가 아닙니다."
                        ]
                    }
                },
                "allOf": [
                    {
                        "properties": {
                            "message": {
                                "type": "string"
                            },
                            "errors": {}
                        }
                    },
                    {
                        "$ref": "#/components/schemas/ResponseData"
                    }
                ]
            },
            "ResponseList": {
                "description": "리스폰 리스트 객체",
                "type": "object",
                "allOf": [
                    {
                        "properties": {
                            "results": {}
                        }
                    },
                    {
                        "$ref": "#/components/schemas/ResponseData"
                    }
                ]
            },
            "ResponseObject": {
                "description": "리스폰 단일 객체",
                "type": "object",
                "allOf": [
                    {
                        "properties": {
                            "result": {}
                        }
                    },
                    {
                        "$ref": "#/components/schemas/ResponseData"
                    }
                ]
            },
            "NewUser": {
                "description": "신규 유저",
                "properties": {
                    "username": {
                        "description": "로그인 네임",
                        "type": "string",
                        "example": "testuser"
                    },
                    "name": {
                        "description": "회원명",
                        "type": "string",
                        "example": "홍길동"
                    },
                    "password": {
                        "description": "패스워드",
                        "type": "string",
                        "example": "qdw12312A3!3"
                    },
                    "nickname": {
                        "description": "별칭",
                        "type": "string",
                        "example": "laraveluser"
                    },
                    "phone_number": {
                        "description": "전화 번호",
                        "type": "string",
                        "example": "01011112222"
                    },
                    "email": {
                        "description": "이메일",
                        "type": "string",
                        "example": "test@example.com"
                    },
                    "gender": {
                        "description": "성별<br>`M`: 남성, `F`: 여성",
                        "type": "string",
                        "enum": [
                            "M",
                            "F"
                        ],
                        "example": "M"
                    }
                },
                "type": "object"
            },
            "UserId": {
                "description": "회원 고유 아이디",
                "type": "integer",
                "format": "int64",
                "example": 1
            },
            "User": {
                "description": "회원 정보",
                "allOf": [
                    {
                        "properties": {
                            "id": {
                                "$ref": "#/components/schemas/UserId"
                            }
                        },
                        "type": "object"
                    },
                    {
                        "$ref": "#/components/schemas/NewUser"
                    }
                ],
                "not": [
                    "password"
                ]
            }
        }
    },
    "tags": [
        {
            "name": "Login",
            "description": "회원 가입 및 토큰 발급"
        }
    ],
    "security": [
        []
    ]
}